ADD bitrixvm backup And remote backup over ssh

+++ dump-databases.sh
@@ -11,26 +11,27 @@ echo "$(date +'%F %T') #################### DUMP MYSQL DATABASES TO CORRESPONDIN
 while read USER ; do
   # Create dir where the user databases will be stored
   DESTINATION=$HOME_DIR/$USER/$DB_DUMP_DIR_NAME
+  # Clean destination
+  rm -rf $DESTINATION
   mkdir -p $DESTINATION
   
   # Get databases
   while read DATABASE ; do
-    
-    if [ -e "/usr/local/bin/mydumper" ] || [ -e "/usr/bin/mydumper" ]; 
-      then
-        mkdir -p $DESTINATION/$DATABASE
-        # Clean destination
-        rm -f $DESTINATION/$DATABASE/*
-        mydumper -v 1 --triggers --events --routines --no-locks --less-locking --rows=1000 --threads=1 --compress --database=$DATABASE --outputdir=$DESTINATION/$DATABASE
-      else
-        # Clean destination
-        rm -f $DESTINATION/*
-        mysqldump $DATABASE --opt --routines | gzip > $DESTINATION/$DATABASE.sql.gz
-        echo "$(date +'%F %T') -- $DATABASE > $DESTINATION/$DATABASE.sql.gz"
+    if [ ! -z "$DATABASE" ] ; then
+      if [ -e "/usr/local/bin/mydumper" ] || [ -e "/usr/bin/mydumper" ]; 
+        then
+          mkdir -p $DESTINATION/$DATABASE
+          mydumper -v 1 --triggers --events --routines --no-locks --less-locking --rows=1000 --threads=1 --compress --database=$DATABASE --outputdir=$DESTINATION/$DATABASE
+        else
+          # Clean destination
+          rm -f $DESTINATION/*
+          mysqldump $DATABASE --opt --routines | gzip > $DESTINATION/$DATABASE.sql.gz
+          echo "$(date +'%F %T') -- $DATABASE > $DESTINATION/$DATABASE.sql.gz"
+      fi
+      # Fix permissions
+      chown -R $USER:$USER $DESTINATION
+      let DB_COUNT++
     fi
-    # Fix permissions
-    chown -R $USER:$USER $DESTINATION
-    let DB_COUNT++
   done < <(v-list-databases $USER | cut -d " " -f1 | awk '{if(NR>2)print}')
 done < <(v-list-users | cut -d " " -f1 | awk '{if(NR>2)print}')
